conf/requirements/install.txt
certifi==2019.9.11
chardet==3.0.4
cx-Oracle==7.3.0
defusedxml==0.6.0
Django==2.0.6
django-allauth==0.40.0
django-cors-headers==3.2.0
django-mysql==3.3.0
django-suit==0.2.28
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.8
mysql-connector-python==8.0.18
oauthlib==3.1.0
protobuf==3.10.0
PyJWT==1.7.1
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
supervisor==4.1.0
urllib3==1.25.7
mysqlclient==1.4.6
